+++ app/assets/javascripts/folio/input.js
@@ -8,3 +8,4 @@
 //= require folio/input/phone
 //= require folio/input/redactor
 //= require folio/input/remote_autocomplete
+//= require folio/input/tags