+++ test/controllers/folio/console/api/tags_controller_test.rb
@@ -6,14 +6,14 @@ class Folio::Console::Api::TagsControllerTest < Folio::Console::BaseControllerTe
   test "react_select" do
     page = create(:folio_page)
 
-    get url_for([:react_select, :console, :api, :tags])
+    get url_for([:console, :api, :tags])
     assert_response :success
     assert_equal([], response.parsed_body["data"])
 
     one = ActsAsTaggableOn::Tag.create!(name: "one")
     two = ActsAsTaggableOn::Tag.create!(name: "two")
 
-    get url_for([:react_select, :console, :api, :tags])
+    get url_for([:console, :api, :tags])
     assert_response :success
     assert_equal([], response.parsed_body["data"])
 
@@ -21,11 +21,11 @@ class Folio::Console::Api::TagsControllerTest < Folio::Console::BaseControllerTe
                                       tag: one,
                                       context: "tags")
 
-    get url_for([:react_select, :console, :api, :tags])
+    get url_for([:console, :api, :tags])
     assert_response :success
     assert_equal(["one"], response.parsed_body["data"])
 
-    get url_for([:react_select, :console, :api, :tags, context: "custom"])
+    get url_for([:console, :api, :tags, context: "custom"])
     assert_response :success
     assert_equal([], response.parsed_body["data"])
 
@@ -33,23 +33,23 @@ class Folio::Console::Api::TagsControllerTest < Folio::Console::BaseControllerTe
                                       tag: two,
                                       context: "custom")
 
-    get url_for([:react_select, :console, :api, :tags, context: "custom"])
+    get url_for([:console, :api, :tags, context: "custom"])
     assert_response :success
     assert_equal(["two"], response.parsed_body["data"])
 
-    get url_for([:react_select, :console, :api, :tags, context: "custom", q: "tw"])
+    get url_for([:console, :api, :tags, context: "custom", q: "tw"])
     assert_response :success
     assert_equal(["two"], response.parsed_body["data"])
 
-    get url_for([:react_select, :console, :api, :tags, context: "custom", q: "foobarbaz"])
+    get url_for([:console, :api, :tags, context: "custom", q: "foobarbaz"])
     assert_response :success
     assert_equal([], response.parsed_body["data"])
 
-    get url_for([:react_select, :console, :api, :tags, q: "o"])
+    get url_for([:console, :api, :tags, q: "o"])
     assert_response :success
     assert_equal(["one"], response.parsed_body["data"])
 
-    get url_for([:react_select, :console, :api, :tags, q: "foobarbaz"])
+    get url_for([:console, :api, :tags, q: "foobarbaz"])
     assert_response :success
     assert_equal([], response.parsed_body["data"])
   end